+++ Changelog.md
@@ -20,6 +20,8 @@
 - **BREALING CHANGE** Change register user response ({"success":true, "ID": 1,...}) to {"success":true, "data":{"id":..}}
 - Fix register, authenticate, and reset password special characters for passwords
 - Replace variables in reset password email subject
+- Add option for sending default WordPress email on user registration
+- Add option for sending default password change WordPRess email when change password endpoint is called
 
 ## 3.6.5 (14 Mar 2026)
 - Fix  CVE-2025-58648 - Stored Cross-Site Scripting vulnerability[PR](https://github.com/nicumicle/simple-jwt-login/pull/162)